+++ packages/shared/src/index.ts
@@ -1,10 +1,10 @@
+export * from './adapters';
 export * from './contracts';
 export * from './porter';
 export * from './schemas';
-export * from './taco-interfaces';
+export * from './taco-signer';
 export type * from './types';
 export * from './utils';
-export * from './viem/';
 export * from './web3';
 
 // Re-exports
