+++ packages/shared/src/taco-signer.ts
@@ -5,10 +5,7 @@
  * - `getAddress()`: Get the signer's address
  * - `signMessage()`: Sign a message (string or bytes)
  *
- * Implementations:
- * - `ViemSignerAdapter`: Adapts viem Account objects to this interface
- * - `ethers.Signer`: Directly compatible via structural typing (because it has the same method signatures)
- * - Future signer adapters can implement this same minimal interface
+ * Future signer adapters can implement this same minimal interface
  */
 export interface TACoSigner {
   /**
