--- packages/shared/src/viem/index.ts
@@ -1,4 +1,0 @@
-export * from './ethers-adapter';
-export * from './signer-adapter';
-export * from './type-guards';
-export * from './types';
